back to Int

+++ dataland-api-key-manager/src/main/kotlin/org/dataland/datalandapikeymanager/api/ApiKeyAPI.kt
@@ -42,7 +42,7 @@ interface ApiKeyAPI {
     @SecurityRequirement(name = "default-bearer-auth")
     @SecurityRequirement(name = "default-oauth")
     fun generateApiKey(
-        @RequestParam(required = false) daysValid: Long?,
+        @RequestParam(required = false) daysValid: Int?,
     ): ResponseEntity<ApiKeyAndMetaInfo>
 
     /** A method to get meta information about the API key status of a specific user, like if that user
